Added atoms and molecules components
