+++ src/components/molecule/Atom.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Atom = props => {
+	return (
+		<span className='atom' style={{ backgroundColor: props.color }}></span>
+	);
+};
+
+export default Atom;
