--- src/components/sphere/types/DuoSphere.js
@@ -1,12 +1,0 @@
-import React, { Fragment } from 'react';
-
-const DuoSphere = () => {
-	return (
-		<Fragment>
-			<span className='atom'></span>
-			<span className='atom'></span>
-		</Fragment>
-	);
-};
-
-export default DuoSphere;
