--- src/components/sphere/types/MonoSphere.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const MonoSphere = () => {
-	return <span className='atom'></span>;
-};
-
-export default MonoSphere;
