--- src/components/sphere/types/TrioSphere.js
@@ -1,13 +1,0 @@
-import React, { Fragment } from 'react';
-
-const TrioSphere = () => {
-	return (
-		<Fragment>
-			<span className='atom'></span>
-			<span className='atom'></span>
-			<span className='atom'></span>
-		</Fragment>
-	);
-};
-
-export default TrioSphere;
